Update AttendanceLog.md
Completed CS 6750: Human-Computer Interaction with an A. 
Mostly completed the veterinary clinic assignment from my Java Object Oriented Programming course but decided to move on because of the poorly written instructions in the the Java Object Oriented Programming course. I learned how to use the Scanner class to take in .txt and .csv files as well as write to them. I learned and applied both handled and unhandled exceptions.